watchdog/requirements.txt

requests==2.31.0
docker==6.1.3
schedule==1.2.0
pyyaml>=6.0
